refactor: rename header includes to be prefixed with `errors`
